fix: assamble 후 에러 fix
